added Lambda arg helper
